requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.4
django-allauth==65.10.0
google-auth==2.40.3
gunicorn==23.0.0
idna==3.10
jwt==1.4.0
mysqlclient==2.2.7
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
upgrade-pip==0.1.4
urllib3==2.5.0
